Move climate icons and mode mapping
