+++ src/panels/lovelace/cards/tile/badges/tile-badge-climate.ts
@@ -1,39 +1,10 @@
-import {
-  mdiClockOutline,
-  mdiFan,
-  mdiFire,
-  mdiHeatWave,
-  mdiPower,
-  mdiSnowflake,
-  mdiWaterPercent,
-} from "@mdi/js";
 import { stateColorCss } from "../../../../../common/entity/state_color";
 import {
+  CLIMATE_HVAC_ACTION_ICONS,
+  CLIMATE_HVAC_ACTION_TO_MODE,
   ClimateEntity,
-  HvacAction,
-  HvacMode,
 } from "../../../../../data/climate";
 import { ComputeBadgeFunction } from "./tile-badge";
-
-export const CLIMATE_HVAC_ACTION_ICONS: Record<HvacAction, string> = {
-  cooling: mdiSnowflake,
-  drying: mdiWaterPercent,
-  fan: mdiFan,
-  heating: mdiFire,
-  idle: mdiClockOutline,
-  off: mdiPower,
-  preheating: mdiHeatWave,
-};
-
-export const CLIMATE_HVAC_ACTION_MODE: Record<HvacAction, HvacMode> = {
-  cooling: "cool",
-  drying: "dry",
-  fan: "fan_only",
-  heating: "heat",
-  idle: "off",
-  off: "off",
-  preheating: "heat",
-};
 
 export const computeClimateBadge: ComputeBadgeFunction = (stateObj) => {
   const hvacAction = (stateObj as ClimateEntity).attributes.hvac_action;
@@ -44,6 +15,6 @@ export const computeClimateBadge: ComputeBadgeFunction = (stateObj) => {
 
   return {
     iconPath: CLIMATE_HVAC_ACTION_ICONS[hvacAction],
-    color: stateColorCss(stateObj, CLIMATE_HVAC_ACTION_MODE[hvacAction]),
+    color: stateColorCss(stateObj, CLIMATE_HVAC_ACTION_TO_MODE[hvacAction]),
   };
 };
